experimental cfsr processing
