+++ grib/src/main/resources/META-INF/services/ucar.nc2.IOServiceProvider
@@ -1,2 +1,2 @@
- ucar.nc2.nc2.grib.grib1.Grib1Iosp
- ucar.nc2.nc2.grib.grib2.Grib2Iosp
+ ucar.nc2.grib.collection.Grib1Iosp
+ ucar.nc2.grib.collection.Grib2Iosp
